refactor: set UrlProvider dependency as a property of EmojiDataGenerator
I am doing this refactor since it makes more sense for the Generate function to take in the unicode version. The responsbility of the Generate function is to create emoji data depending on the version number. The responsibility of the URL provider is to just build URLs based on what the generator needs.